[RegressionTest] Add specific file RegressionScene_list for file searching and rename the Regression_test into RegressionScene_test
